method/epitope-cluster-analysis/epitope_cluster_analysis/requirements.txt
biopython==1.67
networkx==2.2
pandas==0.24.2
matplotlib==3.1.1